% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateKmers.R
\name{generateAllKmers}
\alias{generateAllKmers}
\title{generate kmers}
\usage{
generateAllKmers(k, alphabet = c("A", "C", "G", "T"))
}
\arguments{
\item{k}{length}

\item{alphabet}{default DNA bases}
}
\value{
DNAStringSet with all kmers
}
\description{
generate all kmers of given length
}
\details{
generate all kmers of given length
with middle nucleotide...
}
\author{
mvaniterson
importFrom Biostrings DNAStringSet
}

