% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers.R
\name{kmers}
\alias{kmers}
\title{generate kmers}
\usage{
kmers(k, alphabet = c("A", "C", "G", "T"))
}
\arguments{
\item{k}{length}

\item{alphabet}{default DNA bases}
}
\value{
DNAStringSet with all kmers##'
}
\description{
generate all kmers of given length
}
\details{
generate all kmers of given length
with middle nucleotide...
}
\author{
mvaniterson
}

