% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMGW}
\alias{plotMGW}
\title{plot Minor groove width}
\usage{
plotMGW(shape, seq, add = FALSE, ...)
}
\arguments{
\item{shape}{predicted shapes}

\item{seq}{DNAstring containing sequence}

\item{add}{TRUE/FALSE}

\item{...}{optional plotting parameters}
}
\value{
called for its side-effect
}
\description{
plot minor groove width as function of the sequence
}
\details{
Optionally one other sequence can be added
}
\author{
mvaniterson
}

