% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotShape}
\alias{plotShape}
\title{plot shape as function of sequences}
\usage{
plotShape(shape, seq, type = c("mgw", "prot", "roll", "helt"), add = FALSE,
  ...)
}
\arguments{
\item{shape}{predicted shapes}

\item{seq}{DNAstring containing sequence}

\item{type}{"mgw", "prot", "roll" or "helt"}

\item{add}{TRUE/FALSE}

\item{...}{optional plotting parameters}
}
\value{
called for its side-effect
}
\description{
plot shape as function of function
}
\details{
Optionally one other sequence can be added
}
\author{
mvaniterson
}

