% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotShape.R
\name{plotShapes}
\alias{plotShapes}
\title{plot all four DNAshapes}
\usage{
plotShapes(faFile, subset = NULL)
}
\arguments{
\item{faFile}{file containing sequences (fasta-format)}

\item{subset}{vector of indices selected for plotting}
}
\value{
generates a plot
}
\description{
plot all four DNAshape prediction in one figure
}
\author{
mvaniterson
}

