% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predictShapes.R
\name{predictDNAShapes}
\alias{predictDNAShapes}
\title{DNAshape prediction}
\usage{
predictDNAShapes(seqInput, executable = c("DNAshape", "R"))
}
\arguments{
\item{seqInput}{either DNAString or DNAStringSet}

\item{executable}{use the original DNAshape executable or
the R implementation}
}
\value{
shape profile
}
\description{
run the DNAshape prediction
}
\details{
R wrapper around DNAshape prediction
by Tianyin Zhou et al. NAR 2013
}
\author{
mvaniterson
}

