% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prediction.R
\name{read.shape.profiles}
\alias{read.shape.profiles}
\title{read shape profiles}
\usage{
read.shape.profiles(file = NULL, kmerLength = 5)
}
\arguments{
\item{file}{file name query table}

\item{kmerLength}{length of kmer default 5}
}
\value{
shape profiles for all kmers
}
\description{
read shape profiles (query table)
}
\details{
read shape profiles and extends the reverse complements
}
\author{
mvaniterson
}

