% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readShape}
\alias{readShape}
\title{read DNAshape prediction}
\usage{
readShape(file)
}
\arguments{
\item{file}{file containing predicted shape characteristics}
}
\value{
list with DNAshape prediction each element is the predicted shape of a input sequence
}
\description{
read DNAshape prediction
}
\details{
read DNAshape prediction
}
\author{
mvaniterson
}

