% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predictShapes.R
\name{seq2shape}
\alias{seq2shape}
\title{seq2shape}
\usage{
seq2shape(sequence, kmerShapes = NULL, kmerLength = 5)
}
\arguments{
\item{sequence}{character DNA sequence}

\item{kmerShapes}{if missing load's the extended
(reverse complement) pentamer shape profiles}

\item{kmerLength}{default = 5}
}
\value{
shape profile
}
\description{
predict shape for DNA sequence
}
\details{
R implementation that query's the extended
(reverse complement) pentamer shape profiles
}
\author{
mvaniterson
}

