\name{shape.profiles}
\docType{data}
\alias{shape.profiles}
\title{Shape profiles for all pentamers}
\description{Shape profiles based on Molecular Dynamics/Monte-Carlo
  full atom simulation etc.
}
\usage{shape.profiles}
\format{A data.frame containing 1024 rows, for each pentamer
and 4 columns containing the MGW, roll, helix-twist and propellor twist}
\source{Remo Rohs DNAshape}
\references{reference DNAshape
}
\keyword{datasets}