% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2shape.R
\name{shapediff}
\alias{shapediff}
\title{shapediff}
\usage{
shapediff(seq1, seq2, type = c("rmsd", "eucl"))
}
\arguments{
\item{seq1}{DNA sequence 1}

\item{seq2}{DNA sequence 2}

\item{type}{shape difference measure rmsd or eucl.}
}
\value{
shape difference
}
\description{
calculate shape difference
}
\details{
Calculate shape difference between two sequences
}
\author{
mvaniterson
}

