% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotShape.R
\name{plotShape}
\alias{plotShape}
\title{plot single DNAshape}
\usage{
plotShape(shapes, sequences, shapeType = c("MGW", "ProT", "HelT", "Roll"))
}
\arguments{
\item{shapes}{numeric vector containing the structural features}

\item{sequences}{DNAString}

\item{shapeType}{one of MGW, ProT, HelT or Roll}
}
\value{
generates a plot
}
\description{
plot one of the four DNAshape predictions for < 10 sequences of equal length
}
\details{
plot one of the four DNAshape predictions
MWG and ProT are predicted per bp
Roll and HelT are predicted per step
}
\author{
mvaniterson
}

