% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotShapes}
\alias{plotShapes}
\title{plot all four DNAshapes}
\usage{
plotShapes(faFile)
}
\arguments{
\item{faFile}{file containing sequences (fasta-format)}
}
\value{
generates a plot
}
\description{
plot all four DNAshape prediction in one figure
}
\author{
mvaniterson
}

