% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predictDNAShapes}
\alias{predictDNAShapes}
\title{DNAshape prediction}
\usage{
predictDNAShapes(ifile)
}
\arguments{
\item{ifile}{input file containing sequences (fasta-format)}
}
\value{
write DNAshape predictions to separate files
}
\description{
run the DNAshape prediction
}
\details{
R wrapper around DNAshape prediction
by Tianyin Zhou et al. NAR 2013
}
\author{
mvaniterson
}

