% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2shape.R
\name{seq2shape}
\alias{seq2shape}
\title{seq2shape}
\usage{
seq2shape(sequence, kmerLength = 5)
}
\arguments{
\item{sequence}{character DNA sequence
(reverse complement) pentamer shape profiles}

\item{kmerLength}{default = 5}
}
\value{
shape profile
}
\description{
predict shape for DNA sequence
}
\details{
R implementation that query's the extended
(reverse complement) pentamer shape profiles
TODO add fasta-file input
}
\author{
mvaniterson
}

