% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeprofiles.R
\name{store.shape.profiles}
\alias{store.shape.profiles}
\title{store shape profile}
\usage{
store.shape.profiles()
}
\value{
filename Rdata-object
}
\description{
generate Rdata object with kmer shape profiles
}
\details{
generate Rdata object with kmer shape profiles
}
\author{
mvaniterson
}

